split ingress
