+++ pkg/direktiv/grpc-workflow.go
@@ -1,0 +1,287 @@
+package direktiv
+
+import (
+	"context"
+	"fmt"
+
+	hash "github.com/mitchellh/hashstructure/v2"
+	log "github.com/sirupsen/logrus"
+	"github.com/vorteil/direktiv/pkg/ingress"
+	"github.com/vorteil/direktiv/pkg/model"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+	"google.golang.org/protobuf/types/known/timestamppb"
+)
+
+func (is *ingressServer) AddWorkflow(ctx context.Context, in *ingress.AddWorkflowRequest) (*ingress.AddWorkflowResponse, error) {
+
+	var resp ingress.AddWorkflowResponse
+
+	namespace := in.GetNamespace()
+
+	var active bool
+	if in.Active != nil {
+		active = *in.Active
+	}
+
+	var logToEvents string
+	if in.LogToEvents != nil {
+		logToEvents = *in.LogToEvents
+	}
+
+	var workflow model.Workflow
+	document := in.GetWorkflow()
+	err := workflow.Load(document)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "bad workflow definition: %v", err)
+	}
+
+	wf, err := is.wfServer.dbManager.addWorkflow(ctx, namespace, workflow.ID,
+		workflow.Description, active, logToEvents, document, workflow.GetStartDefinition())
+	if err != nil {
+		return nil, grpcDatabaseError(err, "workflow", workflow.ID)
+	}
+
+	is.wfServer.tmManager.deleteTimerByName("", "", fmt.Sprintf("cron:%s", wf.ID.String()))
+	if active {
+		def := workflow.GetStartDefinition()
+		if def.GetType() == model.StartTypeScheduled {
+			scheduled := def.(*model.ScheduledStart)
+			is.wfServer.tmManager.addCron(fmt.Sprintf("cron:%s", wf.ID.String()), wfCron, scheduled.Cron, []byte(wf.ID.String()))
+		}
+	}
+
+	uid := wf.ID.String()
+	revision := int32(wf.Revision)
+
+	log.Debugf("Added workflow %s/%s: %s", namespace, workflow.ID, uid)
+
+	resp.Uid = &uid
+	resp.Id = &wf.Name
+	resp.Revision = &revision
+	resp.Active = &wf.Active
+	resp.CreatedAt = timestamppb.New(wf.Created)
+
+	return &resp, nil
+
+}
+
+func (is *ingressServer) DeleteWorkflow(ctx context.Context, in *ingress.DeleteWorkflowRequest) (*ingress.DeleteWorkflowResponse, error) {
+
+	var (
+		resp ingress.DeleteWorkflowResponse
+	)
+	uid := in.GetUid()
+
+	err := is.wfServer.dbManager.deleteWorkflow(ctx, uid)
+	if err != nil {
+		return nil, grpcDatabaseError(err, "workflow", uid)
+	}
+
+	err = is.wfServer.tmManager.deleteTimerByName("", "", fmt.Sprintf("cron:%s", uid))
+	if err != nil {
+		log.Error(err)
+	}
+
+	log.Debugf("Deleted workflow: %s", uid)
+
+	resp.Uid = &uid
+
+	return &resp, nil
+
+}
+
+func (is *ingressServer) InvokeWorkflow(ctx context.Context, in *ingress.InvokeWorkflowRequest) (*ingress.InvokeWorkflowResponse, error) {
+
+	var resp ingress.InvokeWorkflowResponse
+
+	namespace := in.GetNamespace()
+	workflow := in.GetName()
+	input := in.GetInput()
+
+	inst, err := is.wfServer.engine.PrepareInvoke(ctx, namespace, workflow, input)
+	if err != nil {
+		return nil, grpcDatabaseError(err, "instance", fmt.Sprintf("%s/%s", namespace, workflow))
+	}
+
+	log.Debugf("Invoked workflow %s/%s: %s", namespace, workflow, inst.id)
+
+	resp.InstanceId = &inst.id
+
+	done := make(chan bool)
+	defer close(done)
+
+	// the workflow started, check if we need to wait
+	// wait sends to chan -> sub ready
+	if in.GetWait() {
+		h, _ := hash.Hash(fmt.Sprintf("%s", inst.id), hash.FormatV2, nil)
+		go syncAPIWait(is.wfServer.config.Database.DB, fmt.Sprintf("api:%d", h), done)
+		<-done
+	}
+
+	go inst.start()
+
+	if in.GetWait() {
+		log.Debugf("waiting for response %v", inst.id)
+		<-done
+		log.Debugf("got response %v", inst.id)
+
+		// query results here
+		wfi, err := is.wfServer.dbManager.getWorkflowInstance(ctx, inst.id)
+		if err != nil {
+			return nil, fmt.Errorf("can not fetch instance id %v for wait request: %v", inst.id, err)
+		}
+		resp.Output = []byte(wfi.Output)
+	}
+
+	return &resp, nil
+
+}
+
+func (is *ingressServer) UpdateWorkflow(ctx context.Context, in *ingress.UpdateWorkflowRequest) (*ingress.UpdateWorkflowResponse, error) {
+
+	var resp ingress.UpdateWorkflowResponse
+
+	uid := in.GetUid()
+
+	var workflow model.Workflow
+	document := in.GetWorkflow()
+	err := workflow.Load(document)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "bad workflow definition: %v", err)
+	}
+
+	var checkRevisionVal int
+	var checkRevision *int
+	if in.Revision != nil {
+		checkRevisionVal = int(in.GetRevision())
+		checkRevision = &checkRevisionVal
+	}
+
+	wf, err := is.wfServer.dbManager.updateWorkflow(ctx, uid, checkRevision, workflow.ID,
+		workflow.Description, in.Active, in.LogToEvents, document, workflow.GetStartDefinition())
+	if err != nil {
+		return nil, grpcDatabaseError(err, "workflow", workflow.ID)
+	}
+
+	is.wfServer.tmManager.deleteTimerByName("", "", fmt.Sprintf("cron:%s", wf.ID.String()))
+	if wf.Active {
+		def := workflow.GetStartDefinition()
+		if def.GetType() == model.StartTypeScheduled {
+			scheduled := def.(*model.ScheduledStart)
+			is.wfServer.tmManager.addCron(fmt.Sprintf("cron:%s", wf.ID.String()), wfCron, scheduled.Cron, []byte(wf.ID.String()))
+		}
+	}
+
+	revision := int32(wf.Revision)
+
+	log.Debugf("Updated workflow: %s", uid)
+
+	resp.Uid = &uid
+	resp.Id = &wf.Name
+	resp.Revision = &revision
+	resp.Active = &wf.Active
+	resp.CreatedAt = timestamppb.New(wf.Created)
+
+	return &resp, nil
+
+}
+
+func (is *ingressServer) GetWorkflows(ctx context.Context, in *ingress.GetWorkflowsRequest) (*ingress.GetWorkflowsResponse, error) {
+
+	var resp ingress.GetWorkflowsResponse
+
+	namespace := in.GetNamespace()
+	offset := in.GetOffset()
+	limit := in.GetLimit()
+
+	workflows, err := is.wfServer.dbManager.getWorkflows(ctx, namespace, int(offset), int(limit))
+	if err != nil {
+		return nil, grpcDatabaseError(err, "namespace", "")
+	}
+
+	wfC, err := is.wfServer.dbManager.getWorkflowCount(ctx, namespace, int(offset), int(limit))
+	if err != nil {
+		return nil, grpcDatabaseError(err, "namespace", "")
+	}
+
+	count := int32(wfC)
+
+	resp.Offset = &offset
+	resp.Limit = &limit
+	resp.Total = &count
+
+	for _, wf := range workflows {
+
+		uid := wf.ID.String()
+		revision := int32(wf.Revision)
+
+		resp.Workflows = append(resp.Workflows, &ingress.GetWorkflowsResponse_Workflow{
+			Uid:         &uid,
+			Id:          &wf.Name,
+			Revision:    &revision,
+			Description: &wf.Description,
+			Active:      &wf.Active,
+			CreatedAt:   timestamppb.New(wf.Created),
+			LogToEvents: &wf.LogToEvents,
+		})
+
+	}
+
+	return &resp, nil
+
+}
+
+func (is *ingressServer) GetWorkflowByName(ctx context.Context, in *ingress.GetWorkflowByNameRequest) (*ingress.GetWorkflowByNameResponse, error) {
+
+	var resp ingress.GetWorkflowByNameResponse
+
+	namespace := in.GetNamespace()
+	name := in.GetName()
+
+	wf, err := is.wfServer.dbManager.getWorkflowByName(ctx, namespace, name)
+	if err != nil {
+		return nil, grpcDatabaseError(err, "workflow", fmt.Sprintf("%s/%s", namespace, name))
+	}
+
+	uid := wf.ID.String()
+	revision := int32(wf.Revision)
+
+	resp.Uid = &uid
+	resp.Name = &wf.Name
+	resp.Revision = &revision
+	resp.Active = &wf.Active
+	resp.CreatedAt = timestamppb.New(wf.Created)
+	resp.Description = &wf.Description
+	resp.Workflow = wf.Workflow
+	resp.LogToEvents = &wf.LogToEvents
+
+	return &resp, nil
+
+}
+
+func (is *ingressServer) GetWorkflowByUid(ctx context.Context, in *ingress.GetWorkflowByUidRequest) (*ingress.GetWorkflowByUidResponse, error) {
+
+	var resp ingress.GetWorkflowByUidResponse
+
+	uid := in.GetUid()
+
+	wf, err := is.wfServer.dbManager.getWorkflowByUid(ctx, uid)
+	if err != nil {
+		return nil, grpcDatabaseError(err, "workflow", uid)
+	}
+
+	revision := int32(wf.Revision)
+
+	resp.Uid = &uid
+	resp.Id = &wf.Name
+	resp.Revision = &revision
+	resp.Active = &wf.Active
+	resp.CreatedAt = timestamppb.New(wf.Created)
+	resp.Description = &wf.Description
+	resp.Workflow = wf.Workflow
+	resp.LogToEvents = &wf.LogToEvents
+
+	return &resp, nil
+
+}
